Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

Can you add some details about the script requirements at the top of the file.

From what i can see this depends on the netapp SDK and it requires it to be installed in a specific location.

Patch-set: 3
Label: Code-Review=-1