Update patch set 1
Patch Set 1:

I feel like this quote is a requirement. "Premature optimisation is the root of all evil." Donald Knuth

How often are you running this script?  The more efficient version takes 19.390 total the slower correct program takes 58.573 total. So it takes 3 times as long but gets the correct one.  But it's correct. 

When I run the 2 scripts on the monash-01 zone I also get different output:

With this patch

 Total instances affected in monash-01 zone: 693
 Sent 424 notifications
 ./outage.py -o -z monash-01  9.72s user 0.24s system 15% cpu 1:04.30 total

Without this patch

 Total instances affected in monash-01 zone: 647
 Sent 416 notifications
 ./outage.py -o -z monash-01  12.99s user 0.31s system 33% cpu 39.871 total


I haven't investigated why these are different, but it's not ideal.  In this case the script is only twice as slow :)

Patch-set: 1