Update patch set 1
Patch Set 1: -Code-Review

> I feel like this quote is a requirement. "Premature optimisation is
 > the root of all evil." Donald Knuth

:-P

So leaving aside whether this script should use ip filtering for a moment...

I don't think we should consider filtering to be buggy, currently AFAIK the relevant field is set correctly for all instances that it matters on (it only doesn't get set when either a cell is misconfigured or the instance is in some error state that never got an IP address - nb: we should be regularly looking for and cleaning those up). If there are historical instances out there with broken records they just need to be fixed.

The IP filtering is quite useful in other circumstances as well, mainly when using the nova cli to look at stuff across one's local cell/s - without using the filter you actually cannot do this in one command with nova cli. Another reason why we should make sure it works.

Of course, if there was just a native zone filter in the API then this would all be moot and we'd have already been using it - maybe a BP needed?

Patch-set: 1
Reviewer: Gerrit User 1000082 <1000082@f182a8af-62ca-4ceb-ab81-14f8fec9c6c5>
Label: Code-Review=0