Update patch set 1
Patch Set 1:

Oh crap, i forked the conversation on the list.  Sorry about that.

 > 
 > So leaving aside whether this script should use ip filtering for a
 > moment...
 > 
 > I don't think we should consider filtering to be buggy, currently
 > AFAIK the relevant field is set correctly for all instances that it
 > matters on (it only doesn't get set when either a cell is
 > misconfigured or the instance is in some error state that never got
 > an IP address - nb: we should be regularly looking for and cleaning
 > those up). If there are historical instances out there with broken
 > records they just need to be fixed. 

Yes we should be trying to ensure consistency across all of the cells and we should be doing something about error-ed instances.  Agreed.

 > The IP filtering is quite useful in other circumstances as well,
 > mainly when using the nova cli to look at stuff across one's local
 > cell/s - without using the filter you actually cannot do this in
 > one command with nova cli. Another reason why we should make sure
 > it works.

I think for most cases IP filtering works fine,  but when it comes to reporting outages, better to be safe than sorry.  We don't audit the accessIP field so there is always a chance that there will be a regression in this behavior.

What is the problem that you see is being solved by using IP filtering?  I can probably implement an equivalent feature.  I just don't see the value in the trade off saving 30s vs a higher level of assurance.


 > Of course, if there was just a native zone filter in the API then
 > this would all be moot and we'd have already been using it - maybe
 > a BP needed?

We could at the very least write a NeCTAR specific API extension to support this.  Which isn't totally stupid, this kinda query would help with reporting.  Which is run every 5 min at the moment.

Patch-set: 1