Fix the email notification during expiry
1) Add the email notification when the project turns to 'stopped'.
2) We should keep current behaviour of not sending notification
when the project is starting 'archiving' because we might attempt to
archive resource several times if it is not successful.
3) Remove the notification sending when the project is 'archived'.

Note: 3) is not good enough, because we would not let user know the exact date
of everything is deleted.
Alternatively, we just keep current behaviour of sending notification only
when the project 'archived' is set. For the exceptional case that the project
is hanging in 'archiving' status for 3 monthes, when we set the project
as archived forcefully, we allow extra 2 weeks for the deletion)

Change-Id: I4b4c1f43de8e80d6c303d7d26ddfa8b96046f08a