Don't delete instance in the archiving phase
When archiving resources, current code will delete the instances if the
instance related archives are found. But it will then miss the step to
set it 'archived' when we run instance expiry.
(after deletion of the out of zone instances the expiry will go finished)

Also it generates inconsistency comparing to the normal process,
which create archives but not delete anything in the archiving phase.

Change-Id: I5324d0c256f2ded34d895d0959a163a244004cfb