Added a "--really-all" option for project audits
This includes the disabled projects in the list to be checked.

It turns out that a lot of allocation records in 'A' state have
disabled (typically expired) projects.  This was being hidden.
(The other way to deal with this would be to do extensive
allocation -> project checks in the allocation audit.  That
would be more code.)

Change-Id: I09b45a647f9d1f861572341df5020794240fa0f5