Fix deletion of projects in non-approved state
When the expirer processed a project with an allocation whose
most recent record is not in 'A' state, it switched to the
most recent 'A' record to get the start and end dates, etc.

But if it then needed to do the final deletion, it would send
a delete for the 'A' record ... which would fail because only
the current record may be modified in any way.

This change switches back to the most recent allocatiion record
for the deletion.

Change-Id: I1b7a1d012033d3ac865e1bb7424a9d57f67e3785