Add option to keep existing dates on a force reprovision
In the case where we may want to re-provision a project (for example
we want to reset quotas or re-add the default dns zone) we will also
want to keep the existing start and end dates.

This commit adds a new --keep-dates option to accomodate for this
but only when --force is given and the allocation already has at
least a start date set.

Change-Id: I390f5ac9311b09180a50b9e417021ffa3e3f0da0