Update patch set 3
Patch Set 3: Code-Review-1

The bad news is that this won't work.  The problem is that the nectar-dashboard/rcallocations code does not clear the start_date and end_date when you approve an amendment to an existing allocation.  (Confirmed by looking at the code ... and in the database in dev)

Therefore, this change would lead to the provisioner not setting the new start and end dates following amendments.

Patch-set: 3
Reviewer: Gerrit User 1000157 <1000157@f182a8af-62ca-4ceb-ab81-14f8fec9c6c5>
Label: Code-Review=-1