Set Freshdesk tickets to Closed instead of Resolved
There's a couple of important distinctions between the ticket states
closed and resolved in terms of communication. Closing a ticket:
 * Doesn't send a follow up how-did-we-do survey email
 * Doesn't send the 'we have resolved your issue' email

For our expiry notifications, these auto response emails are out
of context and confusing to users, and they never put a ticket in
in the first place, so switching them to closed should put an end
to it.

Change-Id: I8b071752299ccf9795611e2db97b6c249787526e