Fix archiver errors WRT secgroup_rule quotas
Setting the secgroup_rule quota to zero in an expiring project
cause the final deletion step to fail ... because a non-zero
quota is required by Neutron to list the sec groups to delete.

The "fix" is to set the secgroup_rule quota to 10 rather than 0.
(That is what we were doing manually anyway.)

Change-Id: I96caacc022fdc98f814e3027a698f5b19afb9a4c