Add some audits for cloudkitty
Add audits for cloudkitty flavors, ensure nova extra spec is set
and that all flavors have a cost.

Add audit to compare running instances to last cloudkitty rating

This isn't perfect and won't work for projects that churn instances
alot but is good for projects with a consistant amount of instances.

Change-Id: I8cf07294661f75cc5649f80d22d4529e27348ca7