Fix trove archiver. Only zero ram and volumes quota
Allocations and trove provisioner only affect trove ram and volumes
quota. The archiver should not zero trove backups and instances
quota.
For example, if a project is renewed after reaching restricted state
and all quotas are zeroed, then the provisioner will only update the
trove ram and volumes quota and the user will not be able to launch.
Further, if the default quotas for trove backups and instances have
been customised or overridden for a project then they would also be
lost if zeroed by archiver as they are not reinstated by the
provisioner on renewal.

Change-Id: I53e0f28036f35867dcba649e5b9b83f9d5476119