Limit project and allocation email CCs
Freshdesk will barff if we try to send an email to more
than 49 CCs.  This can happen if a project / allocation has
an excessive number of Members.  But we can't just truncate
the CCs.  Instead we need to construct the list in a way that
perserves the order of the TenantManager and Member role assocations.

Change-Id: Ic3c5f964fcffdbc69851963f4717c822ee5db80d