Only delete clusters on deleted projects
This prevents the edge case that setting a project to disabled for any
reason will allow the clusters to be deleted when audit runs. It is
safer for the delete to be done when the 'deleted' expiry_status is
reached.

This also takes care of deleting clusters from pt after they have gone
through expiry.

Change-Id: I2d037c87aeae3dacba57c66acede35706080cdf9